% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_addLink.R
\name{addLink}
\alias{addLink}
\title{Add a link on a compositional plot}
\usage{
addLink(DIplot, links, Views, samples, variable = NULL, Dims = c(1, 2),
  addLabel = FALSE, labPos = NULL, projColour = "grey",
  latentSize = 0.25)
}
\arguments{
\item{DIplot}{A list with ggplot object where the links are to be added, and data frames with coordinates}

\item{links}{A matrix with two columns containing either feature names or approximate coordinates}
}
\description{
Add a link on a compositional plot
}
