% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_arrayMult.R
\name{arrayMult}
\alias{arrayMult}
\title{Array multiplication}
\usage{
arrayMult(centralMat, outerMat, ncols = ncol(outerMat))
}
\arguments{
\item{centralMat}{an nxp matrix}

\item{outerMat}{an nxd matrix}

\item{ncols}{an integer, the number of columns of outerMat}
}
\value{
an nxpxd matrix, the stacked matrices of centralMat multiplied to every column of outermat
}
\description{
Array multiplication
}
