% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_buildCompMat.R
\name{buildCompMat}
\alias{buildCompMat}
\title{Build the composition matrix for a certain dimension m dimensions}
\usage{
buildCompMat(colMat, paramEsts, latentVar, m, norm = TRUE,
  id = seq_len(m), subtractMax = TRUE)
}
\arguments{
\item{colMat}{The nxp independence model composition matrix}

\item{paramEsts}{The matrix of feature parameter estimates}

\item{latentVar}{The matrix of latent variables}

\item{m}{the required dimension}

\item{norm}{a boolean, should the composition matrix be normalized?}

\item{id}{The vector of dimensions to consider}

\item{subtractMax}{A boolean, should the maximum be substracted from every
composition prior to exponentiation? Recommended for numerical stability}
}
\value{
A matrix with compositions in the rows
}
\description{
Build the composition matrix for a certain dimension m dimensions
}
