% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMarginalOffset.R
\name{buildMarginalOffset}
\alias{buildMarginalOffset}
\title{Build an offset matrix from an marginal model object}
\usage{
buildMarginalOffset(indepModel, invLink)
}
\arguments{
\item{indepModel}{The fitted marginal model, a list}

\item{invLink}{The inverse link function}
}
\value{
an offset matrix of the size of the data
}
\description{
Build an offset matrix from an marginal model object
}
