% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_buildMu.R
\name{buildMu}
\alias{buildMu}
\title{A function to build the mu matrix}
\usage{
buildMu(offSet, latentVar, paramEsts, distribution, paramMatrix = FALSE)
}
\arguments{
\item{offSet}{the offset matrix}

\item{latentVar, paramEsts, distribution}{Latent variables,
parameter estimates and distribution type}

\item{paramMatrix}{A boolean, are feature parameters provided as matrix}
}
\value{
The mean matrix
}
\description{
A function to build the mu matrix
}
