% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_buildMuMargins.R
\name{buildMuMargins}
\alias{buildMuMargins}
\title{Build the marginal mu matrix}
\usage{
buildMuMargins(x, otherMargin, col)
}
\arguments{
\item{x}{The marginal parameters begin estimated}

\item{otherMargin}{The parameters of the other margin}

\item{col}{A logical, are the column parameters being estimated?}
}
\value{
a matrix of means
}
\description{
Build the marginal mu matrix
}
