% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkMeanVarTrend.R
\name{checkMeanVarTrend}
\alias{checkMeanVarTrend}
\title{Quickly check if the mean variance trend provides a good fit}
\usage{
checkMeanVarTrend(data, meanVarFit = "spline", returnTrend = FALSE,
  ...)
}
\arguments{
\item{data}{The n-by-p data matrix}

\item{meanVarFit}{The type of mean variance fit}

\item{returnTrend}{A boolean, should the estimated trend be returned or
only plotted?}

\item{...}{passed on to the estMeanVarTrend() function}
}
\value{
A plot object
}
\description{
Quickly check if the mean variance trend provides a good fit
}
\examples{
data(Zhang)
par(mfrow = c(1,2))
lapply(extractData(list("microbiome" = zhangMicrobio, "metabolome" = zhangMetabo)),
 checkMeanVarTrend)
 par(mfrow = c(1,1))
}
