% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_checkMonotonicity.R
\name{checkMonotonicity}
\alias{checkMonotonicity}
\title{Check for monotonicity in compositional datasets fro given dimensions}
\usage{
checkMonotonicity(modelObj, Dim)
}
\description{
Check for monotonicity in compositional datasets fro given dimensions
}
