% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_convPlot.R
\name{convPlot}
\alias{convPlot}
\title{Plot the convegrence of the different parameter estimates in a line plot}
\usage{
convPlot(model, latent = is.null(View), nVars = Inf, Dim = 1,
  View = NULL, size = 0.125)
}
\arguments{
\item{model}{A fitted modelDI object}

\item{latent}{A boolean, should latent variable trajectory be plotted}

\item{nVars}{An integer, the number of variables to plot. By default all are plotted}

\item{Dim}{The dimension to be plotted}

\item{View}{An integer, the view to be plotted (if latent  = FALSE)}

\item{size}{The line size}
}
\value{
A ggplot object containing the convergence plot
}
\description{
Plot the convegrence of the different parameter estimates in a line plot
}
\examples{
data(hmp2)
microVirDI = compInt(data = list("microbiome" = microPruneVir,
"virome" = virPrune), distributions = c("quasi", "quasi"),
compositional = c(TRUE, TRUE), M = 2)
convPlot(microVirDI)
}
