% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_deriv2LagrangianFeatures.R
\name{deriv2LagrangianFeatures}
\alias{deriv2LagrangianFeatures}
\title{The score function to estimate the latent variables}
\usage{
deriv2LagrangianFeatures(x, data, distribution, offSet, latentVars, numVar,
  paramEstsLower, mm, Jac, meanVarTrend, links, weights, compositional,
  indepModel, ...)
}
\arguments{
\item{...}{arguments to the jacobian function, currently ignored}
}
\value{
A vector of length n, the evaluation of the score functions of the latent variables
}
\description{
The score function to estimate the latent variables
}
