% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_deriv2LagrangianLatentVars.R
\name{deriv2LagrangianLatentVars}
\alias{deriv2LagrangianLatentVars}
\title{The score function to estimate the latent variables}
\usage{
deriv2LagrangianLatentVars(x, data, distributions, offsets, paramEsts,
  paramMats, numVars, latentVarsLower, n, m, Jac, numSets, meanVarTrends,
  links, varPosts, indepModels, compositional, ...)
}
\arguments{
\item{...}{arguments to the jacobian function, currently ignored}
}
\value{
A vector of length n, the evaluation of the score functions of the latent variables
}
\description{
The score function to estimate the latent variables
}
