% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_deriv2LagrangianLatentVarsConstr.R
\name{deriv2LagrangianLatentVarsConstr}
\alias{deriv2LagrangianLatentVarsConstr}
\title{The score function to estimate the latent variables}
\usage{
deriv2LagrangianLatentVarsConstr(x, data, distributions, offsets,
  paramEsts, paramMats, numVars, latentVarsLower, nn, m, Jac, numSets,
  meanVarTrends, links, numCov, covMat, nLambda1s, varPosts, compositional,
  indepModels, ...)
}
\arguments{
\item{...}{arguments to the jacobian function, currently ignored}
}
\value{
A vector of length nn, the evaluation of the score functions of the latent variables
}
\description{
The score function to estimate the latent variables
}
