% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_derivLagrangianFeatures.R
\name{derivLagrangianFeatures}
\alias{derivLagrangianFeatures}
\title{The score function to estimate the feature parameters}
\usage{
derivLagrangianFeatures(x, data, distribution, offSet, latentVars, numVar,
  paramEstsLower, mm, indepModel, meanVarTrend, weights, compositional,
  ...)
}
\arguments{
\item{data}{A list of data matrices}

\item{distribution, link, invLink, compositional, meanVarTrend, offSet, numVar, marginModel}{Characteristics of the view}

\item{latentVars}{A vector of latent variables}

\item{weights}{The normalization weights}

\item{...}{arguments to the jacobian function, currently ignored}
}
\value{
A vector with the evaluation of the score functions of the feature parameters
}
\description{
The score function to estimate the feature parameters
}
