% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_derivLagrangianLatentVars.R
\name{derivLagrangianLatentVars}
\alias{derivLagrangianLatentVars}
\title{The score function to estimate the latent variables}
\usage{
derivLagrangianLatentVars(x, data, distributions, offsets, paramEsts,
  paramMats, numVars, n, m, numSets, meanVarTrends, links, varPosts,
  latentVarsLower, compositional, indepModels, ...)
}
\arguments{
\item{...}{arguments to the jacobian function, currently ignored}
}
\value{
A vector of length n, the evaluation of the score functions of the latent variables
}
\description{
The score function to estimate the latent variables
}
