% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivLagrangianLatentVarsConstr.R
\name{derivLagrangianLatentVarsConstr}
\alias{derivLagrangianLatentVarsConstr}
\title{The score function to estimate the latent variables}
\usage{
derivLagrangianLatentVarsConstr(x, data, distributions, offsets, paramEsts,
  numVars, latentVarsLower, n, m, numSets, meanVarTrends, links, covMat,
  numCov, centMat, nLambda1s, varPosts, compositional, indepModels,
  paramMats, ...)
}
\arguments{
\item{x}{The current estimates of the latent variables}

\item{latentVarsLower}{The parameter estimates of the lower dimensions}

\item{n}{The number of samples}

\item{m}{The dimensions}

\item{numSets}{The number of views}

\item{covMat}{The covariance matrix}

\item{numCov}{The number of covariates}

\item{centMat}{A centering matrix}

\item{nLambda1s}{The number of dummy variables}

\item{compositional, links, indepModels, meanVarTrends, numVars, distributions, data, offsets, varPosts, paramMats, paramEsts}{Lists of information on all the views}

\item{...}{arguments to the jacobian function, currently ignored}
}
\value{
A vector of length n, the evaluation of the score functions of the latent variables
}
\description{
The score function to estimate the latent variables
}
