% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_estFeatureParameters.R
\name{estFeatureParameters}
\alias{estFeatureParameters}
\title{Estimate the feature parameters}
\usage{
estFeatureParameters(paramEsts, lambdasParams, seqSets, data,
  distributions, offsets, nCores, m, JacFeatures, meanVarTrends,
  latentVars, numVars, control, weights, compositional, indepModels, fTol,
  newtonRaphson, allowMissingness, maxItFeat, ...)
}
\arguments{
\item{paramEsts}{Current list of parameter estimates for the different views}

\item{data}{A list of data matrices}

\item{distributions}{A character vector describing the distributions}

\item{offsets}{A list of offset matrices}

\item{latentVars}{A vector of latent variables}

\item{numVars}{The number of variables}

\item{paramEstsLower}{The parameter estimates of the lower dimensions}

\item{n, m}{The number of samples and dimensions, respectively}

\item{Jac}{a jacobian matrix, pre-filled where possible}
}
\value{
A vector of length n, the estimates of the latent variables
}
\description{
Estimate the feature parameters
}
\details{
If forking is available on the OS and the number of cores specified
  is larger than 1, this function mill multithread the estimation of the
  feature parameters for the different views.
}
