% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_estLatentVars.R
\name{estLatentVars}
\alias{estLatentVars}
\title{Estimate the latent variables}
\usage{
estLatentVars(latentVars, lambdasLatent, constrained, fTol, ...)
}
\arguments{
\item{latentVars}{A vector of latent variables}

\item{data}{A list of data matrices}

\item{distributions}{A character vector describing the distributions}

\item{offsets}{A list of offset matrices}

\item{paramEsts}{A list of parameter estimates for the different views}

\item{numVars}{The number of variables}

\item{latentVarsLower}{The latent variables of the lower dimensions}

\item{n, m}{The number of samples and dimensions, respectively}

\item{Jac}{a jacobian matrix, pre-filled where possible}
}
\value{
A vector of length n, the estimates of the latent variables
}
\description{
Estimate the latent variables
}
