% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_estMeanVarTrend.R
\name{estMeanVarTrend}
\alias{estMeanVarTrend}
\title{Estimate a column-wise mean-variance trend}
\usage{
estMeanVarTrend(data, meanMat, baseAbundances, libSizes, plot = FALSE,
  meanVarFit, degree = 2L, compVector = FALSE, constraint = "none",
  ...)
}
\arguments{
\item{data}{the data matrix with n rows}

\item{meanMat}{the estimated mean matrix}

\item{plot}{A boolean, should the trend be plotted?}
}
\value{
A list with components
\item{meanVarTrend}{An smoothed trend function, that can map a mean on a variance}
\item{meanVarTrendDeriv}{A derivative function of this}
}
\description{
Estimate a column-wise mean-variance trend
}
