% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_estOff.R
\name{estOff}
\alias{estOff}
\title{Estimate the column parameters of the independence model}
\usage{
estOff(data, distribution, rowOff, colOff, meanVarTrend, col,
  newtonRaphson, libSizes, ...)
}
\arguments{
\item{data}{a list of data matrices with the same number of samples n in the rows.
Also phyloseq objects are acceptable}

\item{distribution}{a character string describing which distributional assumption should be used.}

\item{rowOff, colOff}{current values for row and column offsets}

\item{meanVarTrend}{The estimated mean-variance trend}

\item{col}{A logical, should column offsets be estimated}

\item{...}{passed onto nleqslv}
}
\description{
Estimate the column parameters of the independence model
}
