% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_extractData.R
\name{extractData}
\alias{extractData}
\title{Helper function to extract data matrix from phyloseq, expressionset objects etc. Also filers out all zero rows}
\usage{
extractData(data, logTransformMicroArray = TRUE)
}
\arguments{
\item{data}{The list of data objects, either matrix, phyloseq or ExpressionSet objects}

\item{logTransformMicroArray}{A boolean, should array data be logtransformed}
}
\value{
the raw data matrices, samples in the rows
}
\description{
Helper function to extract data matrix from phyloseq, expressionset objects etc. Also filers out all zero rows
}
\examples{
data(hmp2)
matrixList = extractData(list("microbiome" = microPruneVir,
"virome" = virPrune))
}
