% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_extractData.R
\docType{methods}
\name{extractMat}
\alias{extractMat}
\alias{extractMat,phyloseq-method}
\alias{extractMat,ExpressionSet-method}
\alias{extractMat,matrix-method}
\title{A function to extract a data matrix from a number of objects}
\usage{
extractMat(Y, ...)

\S4method{extractMat}{phyloseq}(Y, ...)

\S4method{extractMat}{ExpressionSet}(Y, logTransformGaussian, ...)

\S4method{extractMat}{matrix}(Y, ...)
}
\arguments{
\item{Y}{a phyloseq or eSet object, or another object, or a raw data matrix}

\item{...}{additional arguments for the extractor function}

\item{logTransformGaussian}{A boolean, should array data be logtransformed}
}
\value{
A data matrix with samples in the rows and features in the columns
}
\description{
A function to extract a data matrix from a number of objects
}
