% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_filterConfounders.R
\name{filterConfounders}
\alias{filterConfounders}
\title{Filter out the effect of known confounders}
\usage{
filterConfounders(confMat, data, distribution, link, invLink,
  compositional, control, meanVarTrend, offSet, numVar, marginModel,
  biasReduction, allowMissingness, maxItFilt)
}
\description{
Filter out the effect of known confounders
}
