% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInflLatentVar.R
\name{getInflLatentVar}
\alias{getInflLatentVar}
\title{Extract the influence on the estimation of the latent variable}
\usage{
getInflLatentVar(score, InvJac, i)
}
\arguments{
\item{score}{The score matrix}

\item{InvJac}{The inverse Jacobian}

\item{i}{the sample index}
}
\value{
The influence of all observations on the i-th latent variable
}
\description{
Extract the influence on the estimation of the latent variable
}
