% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_data.R
\docType{data}
\name{hmp2samVar}
\alias{hmp2samVar}
\title{Baseline sample variables of IBD patients}
\format{A dataframe with sample variables
\describe{
  \item{hmp2samVar}{Sample variables matching the microbiome and virome datasets}
}}
\source{
\url{https://www.ibdmdb.org/}
}
\usage{
hmp2samVar
}
\description{
Baseline covariates of IBD patients and healthy controls,
}
\keyword{datasets}
