% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_inflPlot.R
\name{inflPlot}
\alias{inflPlot}
\title{A ggplot line plot showing the influences}
\usage{
inflPlot(modelObj, plotType = "pointplot", pointFun = "sum",
  lineSize = 0.07, Dim = 1, ...)
}
\description{
A ggplot line plot showing the influences
}
