% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence.combi.R
\name{influence.combi}
\alias{influence.combi}
\title{Evaluate the influence function}
\usage{
\method{influence}{combi}(modelObj, samples = is.null(View), Dim = 1,
  View = NULL)
}
\arguments{
\item{modelObj}{The model object}

\item{samples}{A boolean, should we look at sample variables? Throws an error otherwise}

\item{Dim, View}{Integers, the dimension and views required}
}
\value{
A list with components
\item{score}{The evaluation of the score function}
\item{InvJac}{The inverted jacobian matrix}
}
\description{
Evaluate the influence function
}
\details{
Especially the influence of the different views on the latent
variable or gradient estimation may be of interest. The influence values are
not all calculated. Rather, the score values and inverse jacobian are returned
so they can easily be calculated
}
