% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_influence.modelDI.R
\name{influence.modelDI}
\alias{influence.modelDI}
\title{Evaluate the influence function}
\usage{
\method{influence}{modelDI}(modelObj, latent = is.null(View), Dim = 1,
  View = NULL)
}
\description{
Evaluate the influence function
}
\details{
Especially the influence of the different views on the latent variable or gradient estimation may be of interest
}
