% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_jacConfounders.R
\name{jacConfounders}
\alias{jacConfounders}
\title{Jacobian when estimating confounder variables}
\usage{
jacConfounders(confMat, data, distribution, x, meanVarTrend, offSet)
}
\description{
Jacobian when estimating confounder variables
}
