% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_jacConfoundersComp.R
\name{jacConfoundersComp}
\alias{jacConfoundersComp}
\title{Jacobian for conditioning under compositionality}
\usage{
jacConfoundersComp(x, confMat, data, meanVarTrend, marginModel,
  allowMissingness, biasReduction, subtractMax = TRUE)
}
\arguments{
\item{x}{Confounder parameter estimates}

\item{confMat}{confounder matrix}

\item{data}{data}

\item{meanVarTrend}{mean variance trend}

\item{allowMissingness}{A boolean, are missing values allowed}
}
\value{
the jacobian matrix
}
\description{
Jacobian for conditioning under compositionality
}
