% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_jacLatentVarsConstr.R
\name{jacLatentVarsConstr}
\alias{jacLatentVarsConstr}
\title{Evaluate the jacobian for estimating the latent variable for one view for constrained ordination}
\usage{
jacLatentVarsConstr(latentVar, data, distribution, paramEsts, offSet,
  meanVarTrend, numCov, covMat, varPosts, compositional, mm, indepModel,
  latentVarsLower, ...)
}
\value{
The jacobian matrix
}
\description{
Evaluate the jacobian for estimating the latent variable for one view for constrained ordination
}
