% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_data.R
\docType{data}
\name{microPruneVir}
\alias{microPruneVir}
\title{Microbiomes of IBD patients and healthy controls}
\format{A phyloseq object containing microbiome data on IBD patients and healthy controls
\describe{
  \item{microPruneVir}{The microbiome dataset pruned for matches with the virome object}
 }}
\source{
\url{https://www.ibdmdb.org/}
}
\usage{
microPruneVir
}
\description{
Microbiome sequencing data of IBD patients and healthy controls,
together with baseline covariates
}
\keyword{datasets}
