% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_modelDI.R
\name{modelDI}
\alias{modelDI}
\title{Estimate the model-based data integration method}
\usage{
modelDI(data, M = 3L, covariates = NULL, distributions, compositional,
  maxIt = 300L, tol = 0.001, verbose = FALSE, prevCutOff = 0.95,
  minFraction = 0.1, logTransformMicroArray = TRUE,
  confounders = NULL, nleq.control = list(maxit = 1000L, cndtol =
  1e-16), record = TRUE, weights = NULL, fTol = 1e-05, nCores = 1,
  meanVarFit = "spline", maxFeats = 2000, dispFreq = 10L,
  allowMissingness = FALSE, biasReduction = TRUE, maxItFeat = 20L,
  maxItFilt = 50L)
}
\arguments{
\item{data}{a list of data matrices with the same number of samples n in the rows.
Also phyloseq objects are acceptable}

\item{M}{the required dimension of the fit, a non-negative integer}

\item{covariates}{a dataframe of n samples with sample-specific variables.}

\item{distributions}{a character vector describing which distributional assumption should be used. See details.}

\item{compositional}{A logical vector with the same length as "data",
indicating if the datasets should be treated as compositional}

\item{maxIt}{an integer, the maximum number of iterations}

\item{tol}{A small scalar, the convergence tolerance}

\item{verbose}{Logical. Should verbose output be printed to the console?}

\item{prevCutOff}{a scalar, the prevalance cutoff for the trimming.}

\item{minFraction}{a scalar, each taxon's total abundance
should equal at least the number of samples n times minFraction,
  otherwise it is trimmed.}

\item{confounders}{A dataframe or a list of dataframes with the same length as data.
In the former case the same dataframe is used for conditioning,
In the latter case each view has its own conditioning variables (or NULL).}

\item{record}{A boolean, should intermediate estimates be stored? Can be useful to check convergence}

\item{fTol}{The tolerance for solving the estimating equations}

\item{nCores}{The number of cores to be used in estimating the feature parameters of each view. See details.}

\item{links}{A character vector with names of link functions}
}
\value{
An object of class modelDI, describing the fit
}
\description{
Estimate the model-based data integration method
}
\details{
Using more than one core is only implemented on Unix systems.
Setting nCores > 1 on Windows will use a single core, with a warning.
When the number of cores specified is larger than the number of views,
nCores is silently set to the number of views.
}
