% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_plot.compInt.R
\name{plot.compInt}
\alias{plot.compInt}
\title{Make triplots or higher dimensional plots of the ordination object}
\usage{
\method{plot}{compInt}(x, ..., Dim = c(1, 2), samDf = NULL,
  samCol = NULL, featNum = 20L, featCols = c("darkblue", "darkgreen",
  "darkred", terrain.colors(5)), manExpFactorTaxa = 0.975,
  featSize = 2.5, crossSize = 4, manExpFactorVar = 0.975,
  varNum = nrow(x$alphas), varSize = 2.5, samColValues = NULL,
  samSize = 1.5, strokeSize = 0.05, warnMonotonicity = FALSE,
  returnCoords = FALSE, squarePlot = TRUE, featAlpha = 0.5)
}
\arguments{
\item{x}{the fit}

\item{...}{additional arguments, currently ignored}

\item{Dim}{the dimensions to be plotted}

\item{samDf}{a dataframe of sample variables}

\item{samCol}{A variable name from samDf used to colour the samples}

\item{featNum, varNum}{The number of features and variables to plot}

\item{featCols}{Colours for the features}

\item{manExpFactorTaxa, manExpFactorVar}{Expansion factors for taxa and
variables, normally calculated natively}

\item{featSize, crossSize, varSize, samSize, strokeSize}{Size parameters for
the feature dots, central cross, variable labels, sample dots and sample
strokes}

\item{samColValues}{Colours for the samples}

\item{warnMonotonicity}{A boolean, should a warning be thrown when the
feature proportions of compositional views do not all vary monotonically
with all latent variables?}

\item{returnCoords}{A boolean, should coordinates be returned, e.g. for use
in thrird party software}

\item{squarePlot}{A boolean, should the axes be square? Strongly recommended}

\item{featAlpha}{Controls the transparacny of the features}
}
\value{
A ggplot object containing the plot
}
\description{
Make triplots or higher dimensional plots of the ordination object
}
\examples{
data(hmp2)
microVirDI = compInt(data = list("microbiome" = microPruneVir,
"virome" = virPrune), distributions = c("quasi", "quasi"),
compositional = c(TRUE, TRUE), verbose = TRUE, nCores = 1, M = 2)
plot(microVirDI)
}
