% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_plot.modelDI.R
\name{plot.modelDI}
\alias{plot.modelDI}
\title{Make triplots or higher dimensional plots of the ordination object}
\usage{
\method{plot}{modelDI}(modelObj, Dim = c(1, 2), samDf = NULL,
  samCol = NULL, featNum = 20L, featCols = c("darkblue", "darkgreen",
  "darkred", terrain.colors(5)), manExpFactorTaxa = 0.975,
  featSize = 2.5, crossSize = 4, manExpFactorVar = 0.975,
  varNum = nrow(modelObj$alphas), varSize = 2.5, samColValues = NULL,
  samSize = 1.5, strokeSize = 0.05, warnMonotonicity = FALSE,
  returnCoords = FALSE, squarePlot = TRUE, featAlpha = 0.5)
}
\description{
Make triplots or higher dimensional plots of the ordination object
}
