% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_polyHorner.R
\name{polyHorner}
\alias{polyHorner}
\title{Horner's method to evaluate a polynomial, copied from the polynom package. the most efficient way}
\usage{
polyHorner(coefs, x)
}
\arguments{
\item{coefs}{the polynomial coefficients}

\item{x}{the input values for the polynomial function}
}
\value{
the evaluated polynomial
}
\description{
Horner's method to evaluate a polynomial, copied from the polynom package. the most efficient way
}
