% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_prepareJacMatComp.R
\name{prepareJacMatComp}
\alias{prepareJacMatComp}
\title{prepare the jacobian for the latent variabels compostional}
\usage{
prepareJacMatComp(mu, paramEsts, CompMat0, meanVarTrend, data, libSizes)
}
\arguments{
\item{libSizes}{}
}
\value{
The empty jacobian matrix with entries maximally filled out
}
\description{
prepare the jacobian for the latent variabels compostional
}
