% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_prepareScoreMat.R
\name{prepareScoreMat}
\alias{prepareScoreMat}
\title{Prepare a helper matrix for score function evaluation under quasi-likelihood}
\usage{
prepareScoreMat(data, mu, meanVarTrend, CompMat, libSizes)
}
\arguments{
\item{meanVarTrend}{}
}
\value{
The helper matrix
}
\description{
Prepare a helper matrix for score function evaluation under quasi-likelihood
}
