% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.combi.R
\name{print.combi}
\alias{print.combi}
\title{print an overview of a fitted combi x}
\usage{
\method{print}{combi}(x)
}
\arguments{
\item{x}{a fitted combi x}

\item{...}{Further arguments, currently ignored}
}
\value{
A ggplot x containing the plot
}
\description{
print an overview of a fitted combi x
}
\examples{
data(Zhang)
\donttest{
#Unconstrained
microMetaboInt = combi(
list("microbiome" = zhangMicrobio, "metabolomics" = zhangMetabo),
distributions = c("quasi", "gaussian"), compositional = c(TRUE, FALSE),
logTransformGaussian = FALSE, verbose = TRUE)
#Constrained
microMetaboIntConstr = combi(
    list("microbiome" = zhangMicrobio, "metabolomics" = zhangMetabo),
    distributions = c("quasi", "gaussian"), compositional = c(TRUE, FALSE),
    logTransformGaussian = FALSE, covariates = zhangMetavars, verbose = TRUE)}
    #Load the fits
load(system.file("extdata", "zhangFits.RData", package = "combi"))
print(microMetaboInt)
print(microMetaboIntConstr)
}
