% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_quasiJacIndep.R
\name{quasiJacIndep}
\alias{quasiJacIndep}
\title{The jacobian for column offset estimation}
\usage{
quasiJacIndep(x, data, otherMargin, meanVarTrend, col, libSizes, ...)
}
\value{
the jacobian matrix
}
\description{
The jacobian for column offset estimation
}
