% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_quasiScoreIndep.R
\name{quasiScoreIndep}
\alias{quasiScoreIndep}
\title{Quasi score equations for column offset parameters of sequence count data}
\usage{
quasiScoreIndep(x, data, otherMargin, meanVarTrend, col, libSizes, ...)
}
\arguments{
\item{x}{the initial guess for the current margin}

\item{data}{the data matrix}

\item{otherMargin}{The other margin}

\item{meanVarTrend}{the function describing the mean-variance trend}

\item{col}{A logical, is the column being estimated?}
}
\value{
the evaluated estimating equation
}
\description{
Quasi score equations for column offset parameters of sequence count data
}
