% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.combi.R
\name{scaleCoords}
\alias{scaleCoords}
\title{A helper function to rescale coordinates}
\usage{
scaleCoords(featCoords, latentData, manExpFactorTaxa, featNum = NULL)
}
\arguments{
\item{featCoords}{the feature coordinates to be rescaled}

\item{latentData}{latent variables}

\item{manExpFactorTaxa}{an expansion factor}

\item{featNum}{the number of features to retain}
}
\value{
The rescaled feature coordinates
}
\description{
A helper function to rescale coordinates
}
