% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_scoreConfounders.R
\name{scoreConfounders}
\alias{scoreConfounders}
\title{Score functions for confounder variables}
\usage{
scoreConfounders(x, data, distribution, offSet, confMat, meanVarTrend,
  allowMissingness)
}
\arguments{
\item{allowMissingness}{}
}
\value{
The evaluation of the estimating equations
}
\description{
Score functions for confounder variables
}
