% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_scoreConfoundersComp.R
\name{scoreConfoundersComp}
\alias{scoreConfoundersComp}
\title{Score equations for conditioning under compositionality}
\usage{
scoreConfoundersComp(x, confMat, data, meanVarTrend, marginModel,
  biasReduction, allowMissingness, subtractMax = TRUE)
}
\arguments{
\item{subtractMax}{}
}
\value{
The evaluation of the estimating equations
}
\description{
Score equations for conditioning under compositionality
}
