% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_scoreFeatureParams.R
\name{scoreFeatureParams}
\alias{scoreFeatureParams}
\title{Evaluate the score functions for the estimation of the feature parameters for a
single dataset}
\usage{
scoreFeatureParams(x, data, distribution, offSet, latentVar, meanVarTrend,
  mm, indepModel, compositional, paramEstsLower, allowMissingness, ...)
}
\arguments{
\item{data}{A list of data matrices}

\item{compositional}{A list of booleans indicating compositionality}

\item{allowMissingness}{A boolean indicating whether missing values are
allowed}

\item{...}{Additional arguments passed on to the score and jacobian functions}
}
\value{
A vector with evaluated score function
}
\description{
Evaluate the score functions for the estimation of the feature parameters for a
single dataset
}
