% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_scoreLatentVars.R
\name{scoreLatentVars}
\alias{scoreLatentVars}
\title{Evaluate the score functions for the estimation of the latent variables for a
single dataset}
\usage{
scoreLatentVars(data, distribution, paramEsts, paramMats, offSet,
  latentVar, meanVarTrend, constrained = FALSE, covMat = NULL,
  varPosts, compositional, indepModel, mm, latentVarsLower,
  allowMissingness, ...)
}
\arguments{
\item{data, distribution, offSet, meanVarTrend, indepModel, varPosts, paramEsts, paramMats, compositional}{#' Characteristics of the views}

\item{latentVar}{the latent variable estimates}

\item{constrained}{a boolean, is this a constrained analysis}

\item{covMat}{a matrix of constraining covariates}

\item{mm}{the current dimension}

\item{latentVarsLower}{the lower dimensional latent variables}

\item{allowMissingness}{a boolean, should missing values be allowed}

\item{...}{additional arguments passed on to score and jacobian functions}
}
\value{
A vector of length n, with evaluated score function
}
\description{
Evaluate the score functions for the estimation of the latent variables for a
single dataset
}
