% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_seq_m.R
\name{seq_m}
\alias{seq_m}
\title{A small auxiliary function for the indices of the lagrange multipliers}
\usage{
seq_m(y, normal = TRUE, nLambda1s = 1)
}
\arguments{
\item{normal}{a logical, is there a normalization restriction?}

\item{nLambda1s}{the number of centering restrictions}

\item{m}{an integer, the current dimension}
}
\value{
a vector containing the ranks of the current lagrangian multipliers
}
\description{
A small auxiliary function for the indices of the lagrange multipliers
}
