% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_summary.combi.R
\name{summary.combi}
\alias{summary.combi}
\title{Show an overview of a fitted combi object}
\usage{
\method{summary}{combi}(object, ...)
}
\arguments{
\item{object}{a fitted combi object}

\item{...}{Further arguments, currently ignored}
}
\value{
A ggplot object containing the plot
}
\description{
Show an overview of a fitted combi object
}
\examples{
data(Zhang)
\donttest{
#Unconstrained
microMetaboInt = combi(
list("microbiome" = zhangMicrobio, "metabolomics" = zhangMetabo),
distributions = c("quasi", "gaussian"), compositional = c(TRUE, FALSE),
logTransformGaussian = FALSE, verbose = TRUE)
#Constrained
microMetaboIntConstr = combi(
    list("microbiome" = zhangMicrobio, "metabolomics" = zhangMetabo),
    distributions = c("quasi", "gaussian"), compositional = c(TRUE, FALSE),
    logTransformGaussian = FALSE, covariates = zhangMetavars, verbose = TRUE)}
    #Load the fits
load(system.file("extdata", "zhangFits.RData", package = "combi"))
summary(microMetaboInt)
summary(microMetaboIntConstr)
}
