% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_data.R
\docType{data}
\name{virPrune}
\alias{virPrune}
\title{Viromes of IBD patients and healthy controls}
\format{A phyloseq object containing virome data on IBD patients and healthy controls
\describe{
  \item{virPrune}{The virome dataset pruned for matches with the microbiome object}
}}
\source{
\url{https://www.ibdmdb.org/}
}
\usage{
virPrune
}
\description{
Virome sequencing data of IBD patients and healthy controls
}
\keyword{datasets}
