% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zhangMetabo}
\alias{zhangMetabo}
\title{Metabolomes of mice that underwent Pulsed Antibiotic Treatment
 (PAT) and controls}
\format{
SummarizedExperiment with metabolome data
\describe{
  \item{zhangMetabo}{The metabolome data as a SummarizedExperiment object}
}
}
\source{
\url{https://www.ibdmdb.org/}
}
\usage{
data(Zhang)
}
\description{
Metabolome of mice that underwent Pulsed Antibiotic Treatment
 (PAT) and controls
}
\keyword{datasets}
