% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zhangMetavars}
\alias{zhangMetavars}
\title{Baseline sample variables of PAT and control mice}
\format{
A dataframe with baseline sample variables
\describe{
  \item{zhangMetavars}{The metadata on the mice}
}
}
\source{
\url{https://www.ibdmdb.org/}
}
\usage{
data(Zhang)
}
\description{
Baseline covariates of PAT mice and healthy controls
}
\keyword{datasets}
