% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_data.R
\docType{data}
\name{zhangMicrobio}
\alias{zhangMicrobio}
\title{Microbiomes of mice that underwent Pulsed Antibiotic Treatment
 (PAT) and controls}
\format{A phyloseq object containing microbiome data
\describe{
  \item{zhangMicrobio}{The microbiome dataset pruned for matches with the metabolome object}
 }}
\source{
\url{https://www.ibdmdb.org/}
}
\usage{
data(Zhang)
}
\description{
Microbiome of mice that underwent Pulsed Antibiotic Treatment
 (PAT) and controls
}
\keyword{datasets}
