% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_addLink.R
\name{addLink}
\alias{addLink}
\title{Add a link on a compositional plot}
\usage{
addLink(DIplot, links, Views, samples, variable = NULL, Dims = c(1, 2),
  addLabel = FALSE, labPos = NULL, projColour = "grey",
  latentSize = 0.25)
}
\arguments{
\item{DIplot}{A list with ggplot object where the links are to be added, and data frames with coordinates}

\item{links}{A matrix with two columns containing either feature names or approximate coordinates}

\item{Views}{Indices or names of the views for which the links should be added}

\item{samples}{Sample names or approximate sample coordinates}

\item{variable}{Name of variable in environmental gradient for which link should be plotted}

\item{Dims}{vector of length 2 referring to the model dimensions}

\item{addLabel}{A boolean, should arrow with label be plotted?}

\item{labPos}{The position of the label}

\item{projColour}{The colour of the projection}

\item{latentSize}{Size of the line from the origin to the latent variable dot}
}
\value{
A ggplot object with the links added
}
\description{
Add a link on a compositional plot
}
\examples{
data(hmp2)
microVirDI = compInt(data = list("microbiome" = microPruneVir,
"virome" = virPrune), distributions = c("quasi", "quasi"),
compositional = c(TRUE, TRUE), verbose = TRUE, nCores = 1, M = 2)
Plot = plot(microVirDI, returnCoords = TRUE) #Store the coordinates
addLink(Plot, links = cbind("354259","12239"), Views = 1, samples = c(-0.2,-0.75))
}
