% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_buildOffsetModel.R
\name{buildOffsetModel}
\alias{buildOffsetModel}
\title{Build a marginal offset matrix given a model}
\usage{
buildOffsetModel(modelObj, View)
}
\arguments{
\item{modelObj}{a modelDI object}

\item{View}{The view for which to build the offset}
}
\value{
The offset matrix
}
\description{
Build a marginal offset matrix given a model
}
