% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkMeanVarTrend.R
\name{checkMeanVarTrend}
\alias{checkMeanVarTrend}
\title{Quickly check if the mean variance trend provides a good fit}
\usage{
checkMeanVarTrend(data, meanVarFit = "spline", returnTrend = FALSE,
  ...)
}
\arguments{
\item{data}{Data in any acceptable format (see details ?combi)}

\item{meanVarFit}{The type of mean variance fit, either "cubic" or "spline"}

\item{returnTrend}{A boolean, should the estimated trend be returned (TRUE) or
only plotted (FALSE)?}

\item{...}{passed on to the estMeanVarTrend() function}
}
\value{
A plot object
}
\description{
Quickly check if the mean variance trend provides a good fit
}
\examples{
data(Zhang)
par(mfrow = c(1,2))
lapply(list("microbiome" = zhangMicrobio, "metabolome" = zhangMetabo),
 checkMeanVarTrend)
 par(mfrow = c(1,1))
}
