% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkMonotonicity.R
\name{checkMonotonicity}
\alias{checkMonotonicity}
\title{Check for monotonicity in compositional datasets fro given dimensions}
\usage{
checkMonotonicity(modelObj, Dim)
}
\arguments{
\item{modelObj}{The combi fit}

\item{Dim}{The dimensions considered}
}
\value{
A boolean matrix indicating monotonicity for every feature
}
\description{
Check for monotonicity in compositional datasets fro given dimensions
}
