% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_deriv2LagrangianFeatures.R
\name{deriv2LagrangianFeatures}
\alias{deriv2LagrangianFeatures}
\title{The score function to estimate the latent variables}
\usage{
deriv2LagrangianFeatures(x, data, distribution, offSet, latentVars, numVar,
  paramEstsLower, mm, Jac, meanVarTrend, weights, compositional,
  indepModel, ...)
}
\arguments{
\item{x}{parameter estimates}

\item{data}{A list of data matrices}

\item{distribution, compositional, meanVarTrend, offSet, numVar}{Characteristics of the view}

\item{latentVars}{A vector of latent variables}

\item{paramEstsLower}{lower dimension estimates}

\item{weights}{The normalization weights}

\item{indepModel}{the independence model}

\item{...}{Additional arguments passed on to the score and jacobian functions}
}
\value{
A vector of length n, the evaluation of the score functions of the latent variables
}
\description{
The score function to estimate the latent variables
}
