% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_deriv2LagrangianLatentVars.R
\name{deriv2LagrangianLatentVars}
\alias{deriv2LagrangianLatentVars}
\title{The jacobian function to estimate the latent variables}
\usage{
deriv2LagrangianLatentVars(x, data, distributions, offsets, paramEsts,
  paramMats, numVars, latentVarsLower, n, m, Jac, numSets, meanVarTrends,
  links, varPosts, indepModels, compositional, ...)
}
\arguments{
\item{x}{The current estimates of the latent variables}

\item{data}{Lists of inforamtion on all the views}

\item{distributions, links, compositional, meanVarTrends, offsets, numVars, numSets, paramMats, paramEsts, varPosts, indepModels}{Characteristics of the views}

\item{latentVarsLower}{The parameter estimates of the lower dimensions}

\item{n, m}{integeres, number of samples and dimensions}

\item{Jac}{an empty jacobian matrix}

\item{...}{arguments to the jacobian function, currently ignored}
}
\value{
A vector of length n, the evaluation of the score functions of the latent variables
}
\description{
The jacobian function to estimate the latent variables
}
