% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_deriv2LagrangianLatentVarsConstr.R
\name{deriv2LagrangianLatentVarsConstr}
\alias{deriv2LagrangianLatentVarsConstr}
\title{The score function to estimate the latent variables}
\usage{
deriv2LagrangianLatentVarsConstr(x, data, distributions, offsets,
  paramEsts, paramMats, numVars, latentVarsLower, nn, m, Jac, numSets,
  meanVarTrends, links, numCov, covMat, nLambda1s, varPosts, compositional,
  indepModels, ...)
}
\arguments{
\item{x}{The current estimates of the latent variables}

\item{data}{Lists of inforamtion on all the views}

\item{distributions}{Lists of inforamtion on all the views}

\item{offsets}{Lists of inforamtion on all the views}

\item{paramEsts}{Lists of inforamtion on all the views}

\item{paramMats}{Lists of inforamtion on all the views}

\item{numVars}{Lists of inforamtion on all the views}

\item{latentVarsLower}{The parameter estimates of the lower dimensions}

\item{m}{The dimensions}

\item{Jac}{an empty jacobian matrix}

\item{numSets}{The number of views}

\item{meanVarTrends}{Lists of inforamtion on all the views}

\item{links}{Lists of inforamtion on all the views}

\item{varPosts}{Lists of inforamtion on all the views}

\item{compositional}{Lists of inforamtion on all the views}

\item{indepModels}{Lists of inforamtion on all the views}

\item{...}{arguments to the jacobian function, currently ignored}
}
\value{
A vector of length nn, the evaluation of the score functions of the latent variables
}
\description{
The score function to estimate the latent variables
}
