% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_deriv2LagrangianLatentVarsConstr.R
\name{deriv2LagrangianLatentVarsConstr}
\alias{deriv2LagrangianLatentVarsConstr}
\title{The score function to estimate the latent variables}
\usage{
deriv2LagrangianLatentVarsConstr(x, data, distributions, offsets,
  paramEsts, paramMats, numVars, latentVarsLower, nn, m, Jac, numSets,
  meanVarTrends, links, numCov, covMat, nLambda1s, varPosts, compositional,
  indepModels, ...)
}
\arguments{
\item{x}{The current estimates of the latent variables}

\item{data}{Lists of inforamtion on all the views}

\item{distributions}{Lists of inforamtion on all the views}

\item{offsets}{Characteristics of the views}

\item{paramEsts}{Characteristics of the views}

\item{paramMats}{Characteristics of the views}

\item{numVars}{Characteristics of the views}

\item{latentVarsLower}{The parameter estimates of the lower dimensions}

\item{m}{integeres, number of samples and dimensions}

\item{Jac}{an empty jacobian matrix}

\item{numSets}{Characteristics of the views}

\item{meanVarTrends}{Characteristics of the views}

\item{links}{Characteristics of the views}

\item{numCov}{The number of covariates}

\item{varPosts}{Characteristics of the views}

\item{indepModels}{Characteristics of the views}

\item{...}{arguments to the jacobian function, currently ignored}

\item{distribution, link, invLink, compositional, meanVarTrend, offSet, numVar, marginModel}{Characteristics of the view}
}
\value{
A vector of length nn, the evaluation of the score functions of the latent variables
}
\description{
The score function to estimate the latent variables
}
