% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriv2LagrangianLatentVarsConstr.R
\name{deriv2LagrangianLatentVarsConstr}
\alias{deriv2LagrangianLatentVarsConstr}
\title{The score function to estimate the latent variables}
\usage{
deriv2LagrangianLatentVarsConstr(
  x,
  data,
  distributions,
  offsets,
  paramEsts,
  paramMats,
  numVars,
  latentVarsLower,
  nn,
  m,
  Jac,
  numSets,
  meanVarTrends,
  links,
  numCov,
  covMat,
  nLambda1s,
  varPosts,
  compositional,
  indepModels,
  ...
)
}
\arguments{
\item{x}{The current estimates of the latent variables}

\item{data}{Lists of inforamtion on all the views}

\item{distributions, links, compositional, meanVarTrends, offsets, numVars, paramMats, paramEsts}{Characteristics of the view}

\item{latentVarsLower}{The parameter estimates of the lower dimensions}

\item{nn}{number of samples}

\item{m}{integeres, number of samples and dimensions}

\item{Jac}{an empty jacobian matrix}

\item{numSets}{Characteristics of the views}

\item{numCov}{The number of covariates}

\item{covMat}{the covariates matrix}

\item{nLambda1s}{The number of centering restrictions}

\item{varPosts}{Characteristics of the views}

\item{indepModels}{Characteristics of the views}

\item{...}{arguments to the jacobian function, currently ignored}
}
\value{
A vector of length nn, the evaluation of the score functions of the latent variables
}
\description{
The score function to estimate the latent variables
}
