% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivLagrangianLatentVars.R
\name{derivLagrangianLatentVars}
\alias{derivLagrangianLatentVars}
\title{The score function to estimate the latent variables}
\usage{
derivLagrangianLatentVars(
  x,
  data,
  distributions,
  offsets,
  paramEsts,
  paramMats,
  numVars,
  n,
  m,
  numSets,
  meanVarTrends,
  links,
  varPosts,
  latentVarsLower,
  compositional,
  indepModels,
  ...
)
}
\arguments{
\item{x}{The current estimates of the latent variables}

\item{n}{The number of samples}

\item{m}{The dimensions}

\item{numSets}{The number of views}

\item{latentVarsLower}{The parameter estimates of the lower dimensions}

\item{compositional, links, indepModels, meanVarTrends, numVars, distributions, data, offsets, varPosts, paramMats, paramEsts}{Lists of inforamtion on all the views}

\item{...}{arguments to the jacobian function, currently ignored}
}
\value{
A vector of length n, the evaluation of the score functions of the latent variables
}
\description{
The score function to estimate the latent variables
}
