% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_derivLagrangianLatentVarsConstr.R
\name{derivLagrangianLatentVarsConstr}
\alias{derivLagrangianLatentVarsConstr}
\title{The score function to estimate the latent variables}
\usage{
derivLagrangianLatentVarsConstr(x, data, distributions, offsets, paramEsts,
  numVars, latentVarsLower, n, m, numSets, meanVarTrends, links, covMat,
  numCov, centMat, nLambda1s, varPosts, compositional, indepModels,
  paramMats, ...)
}
\arguments{
\item{...}{arguments to the jacobian function, currently ignored}
}
\value{
A vector of length n, the evaluation of the score functions of the latent variables
}
\description{
The score function to estimate the latent variables
}
