% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_estLatentVars.R
\name{estLatentVars}
\alias{estLatentVars}
\title{Estimate the latent variables}
\usage{
estLatentVars(latentVars, lambdasLatent, constrained, fTol, ...)
}
\arguments{
\item{latentVars}{A vector of latent variables}

\item{lambdasLatent}{A vector of Lagrange multipliers}

\item{constrained}{A boolean, is the ordination constrained?}

\item{fTol}{The convergence tolerance}

\item{...}{additional arguments passed on to score and jacobian functions}
}
\value{
A vector of length n, the estimates of the latent variables
}
\description{
Estimate the latent variables
}
