% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_filterConfounders.R
\name{filterConfounders}
\alias{filterConfounders}
\title{Filter out the effect of known confounders}
\usage{
filterConfounders(confMat, data, distribution, link, invLink,
  compositional, control, meanVarTrend, offSet, numVar, marginModel,
  biasReduction, allowMissingness, maxItFilt)
}
\arguments{
\item{confMat}{A confounder design matrix}

\item{data}{data matrix}

\item{distribution, link, invLink, compositional, meanVarTrend, offSet, numVar, marginModel}{Characteristics of the view}

\item{control}{A list of control elements to the nleqslv function}

\item{biasReduction}{A boolean, should bias reduction be applied}

\item{allowMissingness}{A boolean, are missing values allowed?}

\item{maxItFilt}{An integer, maximum number of iterations}
}
\value{
Parameter estimates accounting for the effects of the confounders
}
\description{
Filter out the effect of known confounders
}
