% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_GramSchmidtOrth.R
\name{GramSchmidtOrth}
\alias{GramSchmidtOrth}
\title{Gram schimdt orhtogonalize a with respect to b, and normalize}
\usage{
GramSchmidtOrth(a, b, weights = 1, norm = TRUE)
}
\arguments{
\item{a}{the vector to be orthogonalized}

\item{b}{the vector to be orthogonalized to}

\item{weights}{weights vector}

\item{norm}{a boolean, should the result be normalized?}
}
\value{
The orthogonalized vector
}
\description{
Gram schimdt orhtogonalize a with respect to b, and normalize
}
