% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_jacConfounders.R
\name{jacConfounders}
\alias{jacConfounders}
\title{Jacobian when estimating confounder variables}
\usage{
jacConfounders(confMat, data, distribution, x, meanVarTrend, offSet)
}
\arguments{
\item{confMat}{Characteristics of the views}

\item{data}{Characteristics of the views}

\item{distribution, offSet}{distribution and offset of the view}

\item{x}{the parameter estimates}

\item{meanVarTrend}{Characteristics of the views}
}
\value{
the jacobian matrix
}
\description{
Jacobian when estimating confounder variables
}
