% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacConfounders.R
\name{jacConfounders}
\alias{jacConfounders}
\title{Jacobian when estimating confounder variables}
\usage{
jacConfounders(confMat, data, distribution, x, meanVarTrend, offSet,
  CompMat, libSizes, allowMissingness)
}
\arguments{
\item{confMat}{Characteristics of the views}

\item{data}{Characteristics of the views}

\item{distribution, offSet}{distribution and offset of the view}

\item{x}{the parameter estimates}

\item{meanVarTrend}{Characteristics of the views}

\item{libSizes, CompMat}{Library sizes and relative abunance}

\item{allowMissingness}{a boolean, should missing values be allowed}
}
\value{
the jacobian matrix
}
\description{
Jacobian when estimating confounder variables
}
