% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_jacConfoundersComp.R
\name{jacConfoundersComp}
\alias{jacConfoundersComp}
\title{Jacobian for conditioning under compositionality}
\usage{
jacConfoundersComp(x, confMat, data, meanVarTrend, marginModel,
  allowMissingness, biasReduction, subtractMax = TRUE)
}
\description{
Jacobian for conditioning under compositionality
}
