% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacConfoundersComp.R
\name{jacConfoundersComp}
\alias{jacConfoundersComp}
\title{Jacobian for conditioning under compositionality}
\usage{
jacConfoundersComp(x, confMat, data, meanVarTrend, marginModel,
  allowMissingness, biasReduction, subtractMax = TRUE)
}
\arguments{
\item{x}{the parameter estimates}

\item{confMat}{Characteristics of the views}

\item{data}{Characteristics of the views}

\item{meanVarTrend}{Characteristics of the views}

\item{marginModel, biasReduction, subtractMax}{The marginal mode, and booleans
indicating bias reduction and maximum subtraction}

\item{allowMissingness}{a boolean, should missing values be allowed}
}
\value{
the jacobian matrix
}
\description{
Jacobian for conditioning under compositionality
}
