% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_jacFeatures.R
\name{jacFeatures}
\alias{jacFeatures}
\title{Evaluate the jacobian for estimating the feature parameters for one view}
\usage{
jacFeatures(latentVars, data, distribution, paramEsts, meanVarTrend,
  offSet, compositional, indepModel, m, paramEstsLower,
  minNumber = .Machine$double.eps^10, allowMissingness, ...)
}
\value{
The jacobian matrix
}
\description{
Evaluate the jacobian for estimating the feature parameters for one view
}
