% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacFeatures.R
\name{jacFeatures}
\alias{jacFeatures}
\title{Evaluate the jacobian for estimating the feature parameters for one view}
\usage{
jacFeatures(latentVars, data, distribution, paramEsts, meanVarTrend,
  offSet, compositional, indepModel, m, paramEstsLower, allowMissingness,
  ...)
}
\arguments{
\item{latentVars}{A vector of latent variables}

\item{data}{A list of data matrices}

\item{distribution, compositional, meanVarTrend, offSet, paramEsts, paramEstsLower, indepModel}{Characteristics of each view}

\item{m}{dimension}

\item{allowMissingness}{a boolean, are missing values allowed?}

\item{...}{Additional arguments passed on to the score and jacobian functions}
}
\value{
The jacobian matrix
}
\description{
Evaluate the jacobian for estimating the feature parameters for one view
}
