% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_jacLatentVars.R
\name{jacLatentVars}
\alias{jacLatentVars}
\title{Evaluate the jacobian for estimating the latent variable for one view}
\usage{
jacLatentVars(latentVar, data, distribution, paramEsts, paramMats, offSet,
  meanVarTrend, n, varPosts, mm, indepModel, latentVarsLower,
  compositional, allowMissingness, ...)
}
\arguments{
\item{...}{additional arguments passed on to score and jacobian functions}
}
\value{
The jacobian matrix
}
\description{
Evaluate the jacobian for estimating the latent variable for one view
}
