% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacLatentVars.R
\name{jacLatentVars}
\alias{jacLatentVars}
\title{Evaluate the jacobian for estimating the latent variable for one view}
\usage{
jacLatentVars(
  latentVar,
  data,
  distribution,
  paramEsts,
  paramMats,
  offSet,
  meanVarTrend,
  n,
  varPosts,
  mm,
  indepModel,
  latentVarsLower,
  compositional,
  allowMissingness,
  ...
)
}
\arguments{
\item{latentVar}{the latent variable estimates}

\item{distribution, compositional, meanVarTrend, offSet, paramEsts, paramMats, indepModel}{Characteristics of each view}

\item{n}{the number of samples}

\item{mm}{the current dimension}

\item{latentVarsLower}{the lower dimensional latent variables}

\item{allowMissingness}{a boolean, should missing values be allowed}

\item{...}{additional arguments passed on to score and jacobian functions}
}
\value{
The diagonal of the jacobian matrix
}
\description{
Evaluate the jacobian for estimating the latent variable for one view
}
