% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_jacLatentVarsConstr.R
\name{jacLatentVarsConstr}
\alias{jacLatentVarsConstr}
\title{Evaluate the jacobian for estimating the latent variable for one view for constrained ordination}
\usage{
jacLatentVarsConstr(latentVar, data, distribution, paramEsts, offSet,
  meanVarTrend, numCov, covMat, varPosts, compositional, mm, indepModel,
  latentVarsLower, ...)
}
\arguments{
\item{latentVar}{current latent variable estimates}

\item{distribution, compositional, meanVarTrend, offSet, paramEsts, indepModel, varPosts, data}{Characteristics of each view}

\item{numCov}{the number of covariates}

\item{covMat}{the covariates matrix}

\item{mm}{the dimension}

\item{latentVarsLower}{latent variable estimates of lower dimensions}

\item{...}{additional arguments passed on to score and jacobian functions}
}
\value{
The jacobian matrix
}
\description{
Evaluate the jacobian for estimating the latent variable for one view for constrained ordination
}
