% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_predictSpline.R
\name{predictSpline}
\alias{predictSpline}
\title{A custom spline prediction function, extending linearly with a slope such that prediction never drops below first bisectant}
\usage{
predictSpline(fit, newdata, lowerSlope, lowerEval, linX, coefsQuad,
  deriv = 0L, meanVarFit, minFit, new.knots, degree)
}
\description{
A custom spline prediction function, extending linearly with a slope such that prediction never drops below first bisectant
}
