% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_prepareJacMat.R
\name{prepareJacMat}
\alias{prepareJacMat}
\title{prepare the jacobian matrix}
\usage{
prepareJacMat(mu, data, meanVarTrend, CompMat, libSizes)
}
\arguments{
\item{mu}{the mean matrix}

\item{data}{the count matrix}

\item{trendEval}{evaluated mean-variance trend}

\item{derivEval}{evaluated derivative of the mean-variance trend}
}
\value{
the matrix which can be summed over
}
\description{
prepare the jacobian matrix
}
