% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareJacMatComp.R
\name{prepareJacMatComp}
\alias{prepareJacMatComp}
\title{prepare the jacobian for the latent variabels compostional}
\usage{
prepareJacMatComp(mu, paramEsts, CompMat0, meanVarTrend, data, libSizes)
}
\arguments{
\item{mu}{the mean matrix}

\item{paramEsts}{Current parameter estimates}

\item{CompMat0}{The compisition matrix}

\item{meanVarTrend}{The mean variance trend}

\item{data}{the count matrix}

\item{libSizes}{The library sizesv}
}
\value{
The empty jacobian matrix with entries maximally filled out
}
\description{
prepare the jacobian for the latent variabels compostional
}
