% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_prepareScoreMat.R
\name{prepareScoreMat}
\alias{prepareScoreMat}
\title{Prepare a helper matrix for score function evaluation under quasi-likelihood}
\usage{
prepareScoreMat(data, mu, meanVarTrend, CompMat, libSizes)
}
\arguments{
\item{data}{the count matrix}

\item{mu}{the mean matrix}

\item{meanVarTrend}{The mean variance trend}

\item{CompMat}{The compisition matrix}

\item{libSizes}{The library sizes}
}
\value{
The helper matrix
}
\description{
Prepare a helper matrix for score function evaluation under quasi-likelihood
}
