% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasiJacIndep.R
\name{quasiJacIndep}
\alias{quasiJacIndep}
\title{The jacobian for column offset estimation}
\usage{
quasiJacIndep(x, data, otherMargin, meanVarTrend, col, libSizes, ...)
}
\arguments{
\item{x}{the initial guess for the current margin}

\item{data}{the data matrix}

\item{otherMargin}{The other margin}

\item{meanVarTrend}{the function describing the mean-variance trend}

\item{col}{A logical, is the column being estimated?}

\item{libSizes}{The library sizes}

\item{...}{passed on to prepareJacMat}
}
\value{
the jacobian matrix
}
\description{
The jacobian for column offset estimation
}
