% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreConfounders.R
\name{scoreConfounders}
\alias{scoreConfounders}
\title{Score functions for confounder variables}
\usage{
scoreConfounders(
  x,
  data,
  distribution,
  offSet,
  confMat,
  meanVarTrend,
  allowMissingness,
  libSizes,
  CompMat
)
}
\arguments{
\item{x}{the parameter estimates}

\item{data, distribution, offSet, confMat, meanVarTrend}{Characteristics of the views}

\item{allowMissingness}{a boolean, should missing values be allowed}

\item{libSizes, CompMat}{Library sizes and relative abunance}
}
\value{
The evaluation of the estimating equations
}
\description{
Score functions for confounder variables
}
