% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_scoreConfoundersComp.R
\name{scoreConfoundersComp}
\alias{scoreConfoundersComp}
\title{Score equations for conditioning under compositionality}
\usage{
scoreConfoundersComp(x, confMat, data, meanVarTrend, marginModel,
  biasReduction, allowMissingness, subtractMax = TRUE)
}
\arguments{
\item{x}{Confounder parameter estimates}

\item{confMat}{confounder matrix}

\item{data}{data}

\item{meanVarTrend}{mean variance trend}

\item{marginModel}{marginal models}

\item{biasReduction}{A boolean, should a bias reduced estimation be applied?}

\item{allowMissingness}{A boolean, are missing values allowed}

\item{subtractMax}{A boolean, shuold the maximum be subtracted before softmax
transformation? Recommended for numerical stability}
}
\value{
The evaluation of the estimating equations
}
\description{
Score equations for conditioning under compositionality
}
