% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_scoreFeatureParams.R
\name{scoreFeatureParams}
\alias{scoreFeatureParams}
\title{Evaluate the score functions for the estimation of the feature parameters for a
single dataset}
\usage{
scoreFeatureParams(x, data, distribution, offSet, latentVar, meanVarTrend,
  mm, indepModel, compositional, paramEstsLower, allowMissingness, ...)
}
\value{
A vector with evaluated score function
}
\description{
Evaluate the score functions for the estimation of the feature parameters for a
single dataset
}
