% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreFeatureParams.R
\name{scoreFeatureParams}
\alias{scoreFeatureParams}
\title{Evaluate the score functions for the estimation of the feature parameters for a
single dataset}
\usage{
scoreFeatureParams(x, data, distribution, offSet, latentVar, meanVarTrend,
  mm, indepModel, compositional, paramEstsLower, allowMissingness, ...)
}
\arguments{
\item{x}{the parameter estimates}

\item{data, distribution, offSet, meanVarTrend, indepModel, compositional, paramEstsLower}{Characteristics of the views}

\item{latentVar}{the latent variables}

\item{mm}{the dimension}

\item{allowMissingness}{a boolean, should missing values be allowed}

\item{...}{Additional arguments passed on to the score and jacobian functions}
}
\value{
A vector with evaluated score function
}
\description{
Evaluate the score functions for the estimation of the feature parameters for a
single dataset
}
