% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_scoreLatentVars.R
\name{scoreLatentVars}
\alias{scoreLatentVars}
\title{Evaluate the score functions for the estimation of the latent variables for a
single dataset}
\usage{
scoreLatentVars(data, distribution, paramEsts, paramMats, offSet,
  latentVar, meanVarTrend, constrained = FALSE, covMat = NULL,
  varPosts, compositional, indepModel, mm, latentVarsLower,
  allowMissingness, ...)
}
\value{
A vector of length n, with evaluated score function
}
\description{
Evaluate the score functions for the estimation of the latent variables for a
single dataset
}
